+++ api/src/main/java/org/allaymc/api/eventbus/event/server/PlayerJoinEvent.java
@@ -1,5 +1,6 @@
 package org.allaymc.api.eventbus.event.server;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.allaymc.api.annotation.CallerThread;
 import org.allaymc.api.annotation.ThreadType;
@@ -12,11 +13,11 @@ import org.allaymc.api.player.Player;
  * @author daoge_cmd
  */
 @CallerThread(ThreadType.NETWORK)
+@AllArgsConstructor
 public class PlayerJoinEvent extends ServerEvent implements CancellableEvent {
+    /**
+     * The player associated with this event.
+     */
     @Getter
     protected Player player;
-
-    public PlayerJoinEvent(Player player) {
-        this.player = player;
-    }
 }